reduce some compiler warnings on losing precision
git-svn-id: https://svn.r-project.org/R/trunk@89494 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/connections.c
@@ -364,7 +364,7 @@ static int buff_fgetc(Rconnection con)
 
 static double buff_seek(Rconnection con, double where, int origin, int rw)
 {
-    size_t unread_len = con->buff_stored_len - con->buff_pos;
+    double unread_len = (double)(con->buff_stored_len - con->buff_pos);
 
     if (rw == 2) /* write */
 	return con->seek(con, where, origin, rw);
@@ -3363,7 +3363,7 @@ static int raw_fgetc(Rconnection con)
 static double raw_seek(Rconnection con, double where, int origin, int rw)
 {
     Rrawconn this = con->private;
-    double newpos;
+    double newpos; // maybe should be size_t
     size_t oldpos = this->pos;
 
     if(ISNA(where)) return (double) oldpos;
@@ -3374,7 +3374,7 @@ static double raw_seek(Rconnection con, double where, int origin, int rw)
     case 3: newpos = (double) this->nbytes + where; break;
     default: newpos = where;
     }
-    if(newpos < 0 || newpos > this->nbytes)
+    if(newpos < 0 || newpos > (double)this->nbytes)
 	error(_("attempt to seek outside the range of the raw connection"));
     else this->pos = (size_t) newpos;
